+++ src/config.ts
@@ -19,7 +19,8 @@ export type Config = {
     maxTimelineHoursToFetch: number;
     maxInitialTimelineToots: number;
     numTootsInFirstFetch: number;
-    numUserTagsToFetchTootsFor: number;
+    numUserParticipatedTagsToFetchTootsFor: number;
+    numUserParticipatedTagToots: number;
     scoringBatchSize: number;
     staleDataDefaultSeconds: number;
     timelineDecayExponent: number;
@@ -68,7 +69,9 @@ export const DEFAULT_CONFIG: Config = {
     maxInitialTimelineToots: 900,        // useful dev options for faster load
     maxTimelineHoursToFetch: 168,        // Maximum length of time to pull timeline toots for
     numTootsInFirstFetch: 80,            // How many toots to pull in the first fetch
-    numUserTagsToFetchTootsFor: 10,      // Pull toots for this many of the user's most participated tags
+    // TODO: increase this but make the load happen after the initial load
+    numUserParticipatedTagsToFetchTootsFor: 10,      // Pull toots for this many of the user's most participated tags
+    numUserParticipatedTagToots: 150,    // How many total toots to include for the user's most participated tags
     scoringBatchSize: 100,               // How many toots to score at once
     staleDataDefaultSeconds: 10 * 60,    // Default how long to wait before considering data stale
     staleDataSeconds: {                  // Dictionary to configure customized timeouts for different kinds of data